# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.AppointmentResponse:
    properties:
      carId:
        type: string
      clientName:
        type: string
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      notes:
        type: string
      service:
        type: string
      status:
        type: string
      time:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.CreateAppointmentRequest:
    properties:
      carID:
        description: ✅ camelCase
        type: string
      customerID:
        description: ✅ camelCase
        type: string
      employeeID:
        description: ✅ camelCase
        type: string
      notes:
        type: string
      reason:
        type: string
      scheduledAt:
        description: ✅ camelCase
        type: string
      scheduledTime:
        description: ✅ camelCase
        type: string
      serviceType:
        description: ✅ camelCase
        type: string
      status:
        type: string
    type: object
...
  handler.SwaggerMessage:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  handler.UpdateAppointmentRequest:
    properties:
      carId:
        description: ✅ camelCase
        type: string
      customerID:
        type: string
      employeeID:
        type: string
      notes:
        type: string
      reason:
        type: string
      scheduledAt:
        description: ✅ camelCase
        type: string
      scheduledTime:
        description: ✅ camelCase
        type: string
      serviceType:
        description: ✅ camelCase
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API de gestión de taller: autenticación JWT, coches, citas y empleados.'
  title: GonsGarage API
  version: "1.0"
paths:
  /api/v1/appointments:
    get:
      parameters:
      - description: Filtro UUID cliente (staff)
        in: query
        name: customerId
        type: string
      - description: Filtro UUID coche
        in: query
        name: carId
        type: string
      - description: Estado (scheduled, confirmed, ...)
        in: query
        name: status
        type: string
      - description: Límite (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: created_at o scheduled_at
        in: query
        name: sortBy
        type: string
      - description: ASC o DESC
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AppointmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar citas
      tags:
      - appointments
    post:
      consumes:
      - application/json
      parameters:
      - description: carID, scheduledAt (RFC3339), serviceType, etc.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear cita
      tags:
      - appointments
  /api/v1/appointments/{id}:
    delete:
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Sin cuerpo
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar cita
      tags:
      - appointments
    get:
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener cita
      tags:
      - appointments
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID de la cita
        in: path
        name: id
        required: true
        type: string
      - description: Campos a fusionar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar cita
      tags:
      - appointments
...
securityDefinitions:
  BearerAuth:
    description: 'JWT: cabecera Authorization con valor Bearer seguido del token (rutas
      bajo /api/v1 salvo /health).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"